ML/requirements.txt
numpy
pandas
scikit-learn
joblib
pillow
tensorflow[and-cuda]
